requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-filter==2.4.0
django-mysql==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
packaging==20.8
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
tencentcloud-sdk-python==3.0.314
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
